Move some types to the `warp_server_client` crate.
